Add Clippy for More Linting
[Clippy][1] requires a nightly Rust compiler. It's been added as an
optional dependency. You can use it like this:

    cargo build --features "dev"

Clippy didn't have to say much after #8 was merged. It's only suggestion
was to add the `is_empty` method for traits that have a `len` method
(this way, `is_empty` could be optimized to always be `O(1)` and Clippy
will remind you not to use `x.len == 0`).

[1]: https://github.com/Manishearth/rust-clippy